// src/models/BasePrice.model.ts
export type BasePriceListResponse = {
  serviceName: string;
  userType: string;
  serviceType: string;
  guestBaseprice: number;
  price: number;
  convenienceprice: number;
  guestConvenienceprice: number;
  isActive: boolean;
  _id: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
};